MAINTAINERS: Add orphaned mainboards
The mainboards are broken out into individual entries in hopes that it
will be easier for someone to claim ownership than if they were lumped
into a single "Orphaned Mainboards" group.

The theory behind this is that a single mainboard is really the easiest
piece of coreboot to maintain.  Hopefully some less-experienced people
will be interested in stepping up to take over ownership of a mainboard.

Signed-off-by: Martin Roth <gaumless@gmail.com>
Change-Id: I9542b3a7cd87fa8656bc0982c08061e9d0513745
Reviewed-on: https://review.coreboot.org/c/coreboot/+/68706
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Felix Singer <felixsinger@posteo.net>